// src/scripts/interfaces.ts
export interface Choice {
  id?: number;
  customProperties?: Record<string, any>;
  disabled?: boolean;
  active?: boolean;
  elementId?: number;
  groupId?: number;
  keyCode?: number;
  label: string;
  placeholder?: boolean;
  selected?: boolean;
  value: string;
  score?: number;
  choices?: Choice[];
}
// ...
export interface EventMap {
  /**
   * Triggered each time an item is added (programmatically or by the user).
   *
   * **Input types affected:** text, select-one, select-multiple
   *
   * Arguments: id, value, label, groupValue, keyCode
   */
  addItem: CustomEvent<{
    id: number;
    value: string;
    label: string;
    groupValue: string;
    keyCode: number;
  }>;

  /**
   * Triggered each time an item is removed (programmatically or by the user).
   *
   * **Input types affected:** text, select-one, select-multiple
   *
   * Arguments: id, value, label, groupValue
   */
  removeItem: CustomEvent<{
    id: number;
    value: string;
    label: string;
    groupValue: string;
  }>;

  /**
   * Triggered each time an item is highlighted.
   *
   * **Input types affected:** text, select-multiple
   *
   * Arguments: id, value, label, groupValue
   */
  highlightItem: CustomEvent<{
    id: number;
    value: string;
    label: string;
    groupValue: string;
  }>;

  /**
   * Triggered each time an item is unhighlighted.
   *
   * **Input types affected:** text, select-multiple
   *
   * Arguments: id, value, label, groupValue
   */
  unhighlightItem: CustomEvent<{
    id: number;
    value: string;
    label: string;
    groupValue: string;
  }>;

  /**
   * Triggered each time a choice is selected **by a user**, regardless if it changes the value of the input.
   *
   * **Input types affected:** select-one, select-multiple
   *
   * Arguments: choice: Choice
   */
  choice: CustomEvent<{ choice: Choice }>;

  /**
   * Triggered each time an item is added/removed **by a user**.
   *
   * **Input types affected:** text, select-one, select-multiple
   *
   * Arguments: value
   */
  change: CustomEvent<{ value: string }>;

  /**
   * Triggered when a user types into an input to search choices.
   *
   * **Input types affected:** select-one, select-multiple
   *
   * Arguments: value, resultCount
   */
  search: CustomEvent<{ value: string; resultCount: number }>;

  /**
   * Triggered when the dropdown is shown.
   *
   * **Input types affected:** select-one, select-multiple
   *
   * Arguments: -
   */
  showDropdown: CustomEvent<undefined>;

  /**
   * Triggered when the dropdown is hidden.
   *
   * **Input types affected:** select-one, select-multiple
   *
   * Arguments: -
   */
  hideDropdown: CustomEvent<undefined>;

  /**
   * Triggered when a choice from the dropdown is highlighted.
   *
   * Input types affected: select-one, select-multiple
   * Arguments: el is the choice.passedElement that was affected.
   */
  highlightChoice: CustomEvent<{ el: PassedElement }>;
}
// ...
export type ClassNameValue = string | string[];

/** Classes added to HTML generated by  By default classnames follow the BEM notation. */
export interface ClassNames {
  /** @default 'choices' */
  containerOuter: ClassNameValue;
  /** @default 'choices__inner' */
  containerInner: ClassNameValue;
  /** @default 'choices__input' */
  input: ClassNameValue;
  /** @default 'choices__input--cloned' */
  inputCloned: ClassNameValue;
  /** @default 'choices__list' */
  list: ClassNameValue;
  /** @default 'choices__list--multiple' */
  listItems: ClassNameValue;
  /** @default 'choices__list--single' */
  listSingle: ClassNameValue;
  /** @default 'choices__list--dropdown' */
  listDropdown: ClassNameValue;
  /** @default 'choices__item' */
  item: ClassNameValue;
  /** @default 'choices__item--selectable' */
  itemSelectable: ClassNameValue;
  /** @default 'choices__item--disabled' */
  itemDisabled: ClassNameValue;
  /** @default 'choices__item--choice' */
  itemChoice: ClassNameValue;
  /** @default 'choices__placeholder' */
  placeholder: ClassNameValue;
  /** @default 'choices__group' */
  group: ClassNameValue;
  /** @default 'choices__heading' */
  groupHeading: ClassNameValue;
  /** @default 'choices__button' */
  button: ClassNameValue;
  /** @default 'is-active' */
  activeState: ClassNameValue;
  /** @default 'is-focused' */
  focusState: ClassNameValue;
  /** @default 'is-open' */
  openState: ClassNameValue;
  /** @default 'is-disabled' */
  disabledState: ClassNameValue;
  /** @default 'is-highlighted' */
  highlightedState: ClassNameValue;
  /** @default 'is-selected' */
  selectedState: ClassNameValue;
  /** @default 'is-flipped' */
  flippedState: ClassNameValue;
  /** @default 'is-loading' */
  loadingState: ClassNameValue;
  /** @default 'has-no-results' */
  noResults: ClassNameValue;
  /** @default 'has-no-choices' */
  noChoices: ClassNameValue;
}

export interface PassedElement extends HTMLElement {
  classNames: ClassNames;
  element: (HTMLInputElement | HTMLSelectElement) & {
    // Extends HTMLElement addEventListener with Choices events
    addEventListener<K extends keyof EventMap>(
      type: K,
      listener: (
        this: HTMLInputElement | HTMLSelectElement,
        ev: EventMap[K],
      ) => void,
      options?: boolean | AddEventListenerOptions,
    ): void;
  };
  type: 'text' | 'select-one' | 'select-multiple';
  isDisabled: boolean;
  parentInstance: Choices;
}
// ...
export interface Options {
// ...
  position: 'auto' | 'top' | 'bottom';
// ...
}